Handle origin checks consistently (#342)
Adds a helper wrapper around window.addEventListener('message') which does the origin check immediately instead of at various places more nested inside the handlers.

Because the security fix wouldn't work if we accidentally compared `event.origin` to the full URL rather than the origin, we now *only* pass around the origin rather than the full URL, and we add the Sandbox path explicitly when setting up the iframe instead of having the full URL exist elsewhere. This does change the second argument passed to `postMessage` to be the origin instead of the full URL, but this appears to be a correct change as the argument is supposed to be an origin anyway.